requirements.txt
torch>=2.8.0
einops>=0.7.0
transformers>=4.38.0
datasets>=2.17.0
wandb>=0.16.0
accelerate>=0.27.0
numpy
tqdm
pyyaml